refactor: remove unused components, replace imports
